write 32bit SCAX messages, and add some hot/mask registers

+++ NSWConfiguration/ConfigSender.h
@@ -80,7 +80,7 @@ class ConfigSender {
 
     /// High level send/read functions
     void sendTpConfig        (const nsw::TPConfig& tp, bool quiet = false);
-    void sendTpConfigRegister(const nsw::TPConfig& tp, uint8_t address, uint8_t message, bool quiet = false);
+    void sendTpConfigRegister(const nsw::TPConfig& tp, uint8_t address, uint32_t message, bool quiet = false);
     void maskTp              (const nsw::TPConfig& tp, bool sim = false);
     std::vector<uint8_t> readTpConfigRegister(const nsw::TPConfig& tp, uint8_t address);
